Made float_val length reflect on length of mask_optional, reorganized motion lora code
